+++ internal/core/port/product.go
@@ -8,7 +8,7 @@ import (
 
 // ProductRepositoryReader is an interface that wraps all the reading operations for a product. 
 type ProductRepositoryReader interface {
-  FindProductByID(ctx context.Context, id int) (*domain.Product, error)
+  FindProductByID(ctx context.Context, id uint64) (*domain.Product, error)
   FindAllProducts(ctx context.Context) ([]*domain.Product, error)
 }
 
@@ -16,7 +16,7 @@ type ProductRepositoryReader interface {
 type ProductRepositoryWriter interface {
   Create(ctx context.Context, p *domain.Product) error
   Update(ctx context.Context, p *domain.Product) error
-  Delete(ctx context.Context, id int) error
+  Delete(ctx context.Context, id uint64) error
 }
 
 // ProductRepository is an interface that wraps all the reading and writing operations for a product.
@@ -27,9 +27,9 @@ type ProductRepository interface {
 
 // ProductService is an interface that wraps all the operations for a product.
 type ProductService interface {
-  GetByID(ctx context.Context, id int) (*domain.Product, error)
+  GetByID(ctx context.Context, id uint64) (*domain.Product, error)
   GetAll(ctx context.Context) ([]*domain.Product, error)
   Create(ctx context.Context, p *domain.Product) error
-  Inactivate(ctx context.Context, id int) error
-  Activate(ctx context.Context, id int) error
+  Inactivate(ctx context.Context, id uint64) error
+  Activate(ctx context.Context, id uint64) error
 }
